NXP-22534: deprecate NXRuntimeTestCase and replace it
