+++ nuxeo-core/nuxeo-core-api/src/test/java/org/nuxeo/ecm/core/api/TestPropertyModel.java
@@ -269,7 +269,6 @@ public class TestPropertyModel {
         return true;
     }
 
-    // Duplicated from NXRuntimeTestCase
     public static URL getResource(String resource) {
         return Thread.currentThread().getContextClassLoader().getResource(resource);
     }
